Add 'line' vs 'character' mode concept to commander
This is driven by a need in the nrf terminal project where we'd like to
be able to use the terminal in a line mode and is a character mode where
the 'line' mode is the standard behaviour of this plugin and the
'character' mode is where all key presses are passed through
uninterrupted to whatever is backing the xterm instance.

The 'character' mode essentially by-passes all the functionality of the
NrfTerminalCommander but we want it in there to so that we can toggle
seamlessly between the two modes. Without this being supported, we would
have to destroy the terminal instance and restart it without the
NrfTerminalCommander addon but that would lose user information in the
terminal history.