Move 'displayName' back to the bottom
This reverts commit 1358055acb82bdea7c57bea9af4e74bb6c51d665.

It seems like the pipeline is failing:

  https://dev.azure.com/NordicSemiconductor/Wayland/_build/results?buildId=14733&view=results

And it is potentially related to this. I would have thought that each
list item is interpretted as an object with keys and so the order of the
keys doesn't matter but it seems like they need to start with 'task',
'script' or 'bash'. Not sure.